Move controllers package into pkg
